+++ jniLibs/MobileRT/src/main/cpp/MobileRT/Scene.cpp
@@ -32,7 +32,7 @@ int Scene::traceLights(Intersection &intersection, const Ray &ray) const noexcep
     const auto lightsSize(static_cast<unsigned int> (lights_.size()));
 
     for (unsigned int i(0u); i < lightsSize; i++) {
-        const Light &light(*this->lights_[static_cast<uint64_t> (i)]);
+        const Light &light(*this->lights_[static_cast<uint32_t> (i)]);
         if (light.intersectL(intersection, ray)) {
             res = static_cast<int> (i);
         }
@@ -46,7 +46,7 @@ int Scene::trace(Intersection &intersection, const Ray &ray) const noexcept {
     const auto primitivesSize(static_cast<unsigned int> (primitives_.size()));
 
     for (unsigned int i(0u); i < primitivesSize; i++) {
-        const Primitive &primitive(*this->primitives_[static_cast<uint64_t> (i)]);
+        const Primitive &primitive(*this->primitives_[static_cast<uint32_t> (i)]);
         if (primitive.intersect(intersection, ray)) {
             res = static_cast<int> (i);
         }
